move NamespaceAnnotations to kube package (#107)
Retrieving this information with the go client seems better than exec'ing kubectl
