+++ kube/client.go
@@ -11,8 +11,25 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
 )
 
+const (
+	enabledAnnotation               = "kube-applier.io/enabled"
+	dryRunAnnotation                = "kube-applier.io/dry-run"
+	pruneAnnotation                 = "kube-applier.io/prune"
+	pruneClusterResourcesAnnotation = "kube-applier.io/prune-cluster-resources"
+)
+
+// KAAnnotations contains the standard set of annotations on the Namespace
+// resource defining behaviour for that Namespace
+type KAAnnotations struct {
+	Enabled               string
+	DryRun                string
+	Prune                 string
+	PruneClusterResources string
+}
+
 // ClientInterface allows for mocking out the functionality of Client when testing the full process of an apply run.
 type ClientInterface interface {
+	NamespaceAnnotations(namespace string) (KAAnnotations, error)
 	PrunableResourceGVKs() ([]string, []string, error)
 }
 
@@ -41,6 +58,23 @@ func New() (*Client, error) {
 	return &Client{
 		clientset: clientset,
 	}, nil
+}
+
+// NamespaceAnnotations returns string values of kube-applier annotations
+func (c *Client) NamespaceAnnotations(namespace string) (KAAnnotations, error) {
+	kaa := KAAnnotations{}
+
+	ns, err := c.clientset.CoreV1().Namespaces().Get(namespace, metav1.GetOptions{})
+	if err != nil {
+		return kaa, err
+	}
+
+	kaa.Enabled = ns.Annotations[enabledAnnotation]
+	kaa.DryRun = ns.Annotations[dryRunAnnotation]
+	kaa.Prune = ns.Annotations[pruneAnnotation]
+	kaa.PruneClusterResources = ns.Annotations[pruneClusterResourcesAnnotation]
+
+	return kaa, nil
 }
 
 // PrunableResourceGVKs returns cluster and namespaced resources as two slices of
