+++ kubectl/mock_client.go
@@ -47,18 +47,3 @@ func (mr *MockClientInterfaceMockRecorder) Apply(path, namespace, dryRun, kustom
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Apply", reflect.TypeOf((*MockClientInterface)(nil).Apply), path, namespace, dryRun, kustomize, pruneWhitelist)
 }
-
-// NamespaceAnnotations mocks base method
-func (m *MockClientInterface) NamespaceAnnotations(namespace string) (KAAnnotations, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "NamespaceAnnotations", namespace)
-	ret0, _ := ret[0].(KAAnnotations)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// NamespaceAnnotations indicates an expected call of NamespaceAnnotations
-func (mr *MockClientInterfaceMockRecorder) NamespaceAnnotations(namespace interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NamespaceAnnotations", reflect.TypeOf((*MockClientInterface)(nil).NamespaceAnnotations), namespace)
-}
